Move client inside RetroMeet module, rename
