+++ test/objects/retro_meet_client_test.rb
@@ -10,7 +10,7 @@ class RetroMeetClientTest < ActiveSupport::TestCase
                                                       .to_return(webfixture_json_file("login_ok"))
 
     expected_response = "eyJhbGciOiJIUzI1NiJ9.eyJhY2NvdW50X2lkIjoxLCJhdXRoZW50aWNhdGVkX2J5IjpbInBhc3N3b3JkIl19.twCmKESFEIkjABPanrfnlU7mpVmrYsCqGIn_Z9oRJaE"
-    assert_equal expected_response, RetroMeetClient.login(login:, password:)
+    assert_equal expected_response, RetroMeet::Client.login(login:, password:)
     assert_requested(:post, "http://localhost:3000/login")
   end
   test "calls the login method with a bad password and gets an error" do
@@ -19,8 +19,8 @@ class RetroMeetClientTest < ActiveSupport::TestCase
     stub_request(:post, "http://localhost:3000/login").with(headers: { "Content-Type" => "application/json" })
                                                       .to_return(webfixture_json_file("login_bad_password"))
 
-    assert_raises RetroMeetClient::UnauthorizedError do
-      RetroMeetClient.login(login:, password:)
+    assert_raises RetroMeet::Client::UnauthorizedError do
+      RetroMeet::Client.login(login:, password:)
     end
     assert_requested(:post, "http://localhost:3000/login")
   end
@@ -30,8 +30,8 @@ class RetroMeetClientTest < ActiveSupport::TestCase
     stub_request(:post, "http://localhost:3000/login").with(headers: { "Content-Type" => "application/json" })
                                                       .to_return(webfixture_json_file("login_bad_login"))
 
-    assert_raises RetroMeetClient::UnauthorizedError do
-      RetroMeetClient.login(login:, password:)
+    assert_raises RetroMeet::Client::UnauthorizedError do
+      RetroMeet::Client.login(login:, password:)
     end
     assert_requested(:post, "http://localhost:3000/login")
   end
@@ -43,7 +43,7 @@ class RetroMeetClientTest < ActiveSupport::TestCase
                                                                .to_return(webfixture_json_file("create_account_ok"))
 
     expected_response = "eyJhbGciOiJIUzI1NiJ9.eyJhY2NvdW50X2lkIjoxNCwiYXV0aGVudGljYXRlZF9ieSI6WyJhdXRvbG9naW4iXSwiYXV0b2xvZ2luX3R5cGUiOiJjcmVhdGVfYWNjb3VudCJ9.Fo5ZXDASNTcea-V4KlX8JjWy7XQOcPHDFcjET0HeIsw"
-    assert_equal expected_response, RetroMeetClient.create_account(login:, password:, birth_date: "1980-01-01")
+    assert_equal expected_response, RetroMeet::Client.create_account(login:, password:, birth_date: "1980-01-01")
     assert_requested(:post, "http://localhost:3000/create-account")
   end
   test "calls the create_account method with good params but already used login and gets the proper error" do
@@ -52,8 +52,8 @@ class RetroMeetClientTest < ActiveSupport::TestCase
     stub_request(:post, "http://localhost:3000/create-account").with(headers: { "Content-Type" => "application/json" })
                                                                .to_return(webfixture_json_file("create_account_existing"))
 
-    assert_raises RetroMeetClient::LoginAlreadyTakenError do
-      RetroMeetClient.create_account(login:, password:, birth_date: "1980-01-01")
+    assert_raises RetroMeet::Client::LoginAlreadyTakenError do
+      RetroMeet::Client.create_account(login:, password:, birth_date: "1980-01-01")
     end
     assert_requested(:post, "http://localhost:3000/create-account")
   end
@@ -63,8 +63,8 @@ class RetroMeetClientTest < ActiveSupport::TestCase
     stub_request(:post, "http://localhost:3000/create-account").with(headers: { "Content-Type" => "application/json" })
                                                                .to_return(webfixture_json_file("create_account_bad_password"))
 
-    assert_raises RetroMeetClient::BadPasswordError do
-      RetroMeetClient.create_account(login:, password:, birth_date: "1980-01-01")
+    assert_raises RetroMeet::Client::BadPasswordError do
+      RetroMeet::Client.create_account(login:, password:, birth_date: "1980-01-01")
     end
     assert_requested(:post, "http://localhost:3000/create-account")
   end
@@ -74,8 +74,8 @@ class RetroMeetClientTest < ActiveSupport::TestCase
     stub_request(:post, "http://localhost:3000/create-account").with(headers: { "Content-Type" => "application/json" })
                                                                .to_return(webfixture_json_file("create_account_bad_login"))
 
-    assert_raises RetroMeetClient::BadLoginError do
-      RetroMeetClient.create_account(login:, password:, birth_date: "1980-01-01")
+    assert_raises RetroMeet::Client::BadLoginError do
+      RetroMeet::Client.create_account(login:, password:, birth_date: "1980-01-01")
     end
     assert_requested(:post, "http://localhost:3000/create-account")
   end
@@ -85,7 +85,7 @@ class RetroMeetClientTest < ActiveSupport::TestCase
     stub_request(:get, "http://localhost:3000/api/profile/info").with(headers: { "Content-Type" => "application/json", "Authorization" => authorization_header })
                                                                 .to_return(webfixture_json_file("profile_info_good"))
     expected_response = BasicProfileInfo.new("0192e1e9-9e3a-7156-91c3-6e38966821eb", "bob tables", DateTime.new(2024, 9, 23, 15, 45, 24, "+02:00"))
-    assert_equal expected_response, RetroMeetClient.new(authorization_header).basic_profile_info
+    assert_equal expected_response, RetroMeet::Client.new(authorization_header).basic_profile_info
     assert_requested(:get, "http://localhost:3000/api/profile/info")
   end
 end
